Added Diorite Utils - Annotations, moved are usages of old annotations to new package.
