+++ diorite-utils/annotations/src/main/java/org/diorite/annotations/package-info.java
@@ -22,5 +22,8 @@
  * SOFTWARE.
  */
 
+/**
+ * Root package of diorite annotation module.
+ */
 @NonnullByDefault
-package org.diorite.commons.annotations;
+package org.diorite.annotations;
