+++ diorite-utils/injection/src/main/java/org/diorite/inject/package-info.java
@@ -28,4 +28,4 @@
 @NonnullByDefault
 package org.diorite.inject;
 
-import org.diorite.commons.annotations.NonnullByDefault;
+import org.diorite.annotations.NonnullByDefault;
